refactor(types): collapsible
