+++ src/lib/builders/collapsible/types.ts
@@ -1,0 +1,8 @@
+import type { createCollapsible } from "./create";
+
+export type CreateCollapsibleArgs = {
+	open?: boolean;
+	disabled?: boolean;
+};
+
+export type CreateCollapsibleReturn = ReturnType<typeof createCollapsible>
